add logfile for output of parallel
The --joblog option for parallel is used to determine the exit status of
the single latexmk run. The joblog is then filtered with awk for any
exitstatus that is not zero, which is then written to an logfile in the
main directory.
When checking this logfile in the main directory the filename is shown
and can be checked.